client API: deprecate MPV_FORMAT_FLAG in favor of MPV_FORMAT_BOOL
With the recent internal changes that moved all OPT_FLAG types to
OPT_BOOL, it makes sense to mirror this in the client API as well.
Instead of MPV_FORMAT_FLAG, we can use MPV_FORMAT_BOOL which works
essentially exactly the same way. Deprecate the former in favor of the
latter. The next commit will update mpv's internal usage.