context_display: reuse VT switcher from drm
Previously using this backend and switching VTs was a guaranteed freeze
that required you to reboot. It's because there can be only one DRM
master at a time and you're supposed to drop/release it when you receive
the right signals. The drm backend was already doing this, but it was
never implemented in the displayvk backend. Technically speaking,
VkDisplayKHR is not drm-specific, but this already has DRM guarded stuff
and uses libdrm when appropriate. The VT switching stuff doesn't have
anything directly to do with libdrm, but it only makes sense there
anyway so we'll implement this within the HAVE_DRM sections.