f_hwtransfer: also consider supported hwuploads in select_format
With the previous change, it is possible that the mp_hwdec_ctx contains
a list of known supported hwuploads. Take this into account when
selecting the format. If the attempted hwupload is known to not work,
use hw_output_fmt as the hw_input_fmt so it forces a software
conversion.