vo_wlshm: support more formats besides XRGB8888
It is required for all compositors to support XRGB8888, so not
surprisingly this was originally implemented with that format hardcoded.
But wl_shm formats are defined in a way to be compatibile with any drm
format, so we can actually support much more if the compositor supports
the formats. Use the format event in wlshm, do some mapping and then
pick a better format if possible.